move save ml opts function
